requirements.txt
playwright
beautifulsoup4
supabase
python-dotenv